Conflicts:
	csrc/ccnd/dir.mk
	csrc/ccnr/dir.mk
	csrc/lib/dir.mk

+++ javasrc/src/main/org/ccnx/ccn/impl/CCNSegmenter.java
@@ -1,7 +1,7 @@
 /*
  * Part of the CCNx Java Library.
  *
- * Copyright (C) 2008-2012 Palo Alto Research Center, Inc.
+ * Copyright (C) 2008-2013 Palo Alto Research Center, Inc.
  *
  * This library is free software; you can redistribute it and/or modify it
  * under the terms of the GNU Lesser General Public License version 2.1
@@ -918,7 +918,7 @@ public class CCNSegmenter {
 	 * Set the timeout on the contained flow controller.
 	 * @param timeout
 	 */
-	public void setTimeout(int timeout) {
+	public void setTimeout(long timeout) {
 		getFlowControl().setTimeout(timeout);
 	}
